hardware.xml: Split into sections for comedi.org and Linux kernel drivers
Use separate sub-sections to list the devices supported by the original
comedi.org drivers and the devices supported in the Linux kernel
sources.

The devices supported by the Linux kernel sources are listed in
devices-linux.xml, built from devices-linux.txt.

Like devices.txt, the information in devices-linux.txt is derived from
special comments in the driver sources, except that devices-linux.txt is
based on the Linux kernel sources, rather than the comedi.org driver
sources.  The scripts to extract this information are 'dump_doc' and
'doc_devlist' in the main 'comedi' sources.  (See the Makefile.am file
in the main 'comedi' sources for how to run them.)